requirements.txt
biopython==1.70
numpy==1.13.3
py==1.5.2
pysam==0.14
pytest==3.2.5
numpy==1.15.1
